refactor(RestController/RaceView): 컨트롤러,뷰, 그리고 서비스의 책임에 맞게 구조 개선
