+++ docs/README.md
@@ -61,6 +61,7 @@
 
 - [ ] Refactoring
     - [x] 일급컬렉션 Cars 도입
+    - [x] 컨트롤러,뷰, 그리고 서비스의 책임에 맞게 구조 개선
 
 ## 3. 순서도
